backend/requirements.txt
flask-sqlalchemy
pymysql
cryptography
flask
flask-cors
jieba
numpy==1.24.3
gensim==4.3.1
pandas
opencc
pypinyin
bertopic>=0.16.0
sentence-transformers>=3.0.0
scikit-learn>=1.0.0
accelerate>=0.20.0
